requirements.txt
aiohttp==3.10.9
py-cord==2.6.1
Flask==3.0.3
meta_ai_api==1.2.1
python-dotenv==1.0.1
streamlit==1.39.0
